CodingKeys helpers
We've sometimes needed to just create CodingKeys on the fly, so pull out the DynamicCodingKey struct I've been using for that and properly document and test it. At the same time, make an EnumCodingKey<Value> version that does the same thing, but with a closer wrapping of an underlying `enum`.

Fix `testCodable` to hopefully not leave Xcode spinning when trying to figure out where it failed too.