Rework tagged enums
The schema uses an odd mix of full strings ("arcana") and short tags ("XY") for tag-like sets.

I wanted to preserve at the Swift layer these enums should be directly useful, e.g. DamageType.psychic and Treasure.arcana, and that the string description could be used for rendering in the UI.

But at the same time, we need to be able to code these into the schema; with the caveat that sometimes an enum might use its full description, and other times a short tag.

This is made harder by Swift's Codable conformance being per-type, unless you wrap it in another type that has its own Codable conformance and skips that.

That turns out to be the best way to handle this, we create a Tagged<Enum> wrapper that can be used when we're already wrapping the conformance. For the top-level cases (where I wanted to avoid that), create a TagSet<Enum> that wraps Set<> but encodes using their tag equivalents.